Main Files
Authentication manager should be put on an object in a boot strapping scene / log in scene. 
auto save list is a simple override of the built-in list class meant to make is so that every time you add or remove a file, it automatically saves the list. 
to auto- save modifications to an item in a list, since the Set function is only called when the list it's self is being assigned, there is a specialized function in the save-load manager that is meant to be called by the list item (example, DragonNPC scriptable) .